server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
websockets>=13.0.0
groq>=0.14.0
python-dotenv==1.0.1
pydantic==2.12.5
# ── AI Teacher backend additions ──────────────────────────────────────────
pydantic-settings==2.13.1
redis==5.0.8
strands-agents>=0.1.0

# ── Payments backend additions ─────────────────────────────────────────────
razorpay==1.4.2
firebase-admin==6.6.0
email-validator==2.2.0
httpx>=0.28.1

# ── LiteLLM Proxy Integration ────────────────────────────────────────────
litellm>=1.79.3
openai>=1.3.0
psycopg2-binary>=2.9.9